fixed new vrouter rpm dependencies for lp/hostname format change in RHEL 7.4
Before rhel 7.4 the hostname in /etc/hostname was the fqdn (long name).
Contrail services use this string to register themselves in config.
When a service is provisioned using the provisioning scripts, the same
string must be used. Therefore puppet used ::fqdn.
With 7.4 the hostname is the shortname. Puppet could use ::hostname but this
would break compatibility with <7.4. Therefore we do a uname -n in an inline
template as this always matches the string in /etc/hostname

Closes Bug: 1704601

Change-Id: I1d7f7017b644cdd48c478ad87cc6eea2614a003f